+++ src/components/ArticleForm.jsx
@@ -1,36 +1,13 @@
 import classNames from "classnames";
-import { useForm } from "react-hook-form";
-import { postArticle } from "../helpers/fetchArticle";
-import { useState } from 'react';
 import ImageUpload from "../helpers/ImageUpload";
+import { useOnSubmitArticle } from "./hooks/useOnSubmitArticles";
 
 const styleLabel = "font-medium text-sm py-1 ";
 const styleInput = "w-full h-[40px] px-2 text-slate-400 text-xs my-1 rounded-lg border bg-gray-100";
 
 export default function ArticleForm() {
-  const { register, handleSubmit, reset } = useForm();
 
-  const onSubmit = handleSubmit(async (data) => {
-    const formData = new FormData();
-    formData.append("title", data.title);
-    formData.append("description", data.description);
-    formData.append("image", imageFile); // Assuming the field name for file is "image"
-
-    try {
-      await postArticle(formData);
-      // Assuming postMainArticle returns the URL of the uploaded image
-      reset();
-    } catch (error) {
-      console.error("Error:", error);
-    }
-    reset();
-  });
-
-  const [imageFile, setImageFile] = useState(null);
-
-  const handleFileChange = (file) => {
-    setImageFile(file);
-  };
+  const { handleFileChange, onSubmit, register } = useOnSubmitArticle()
 
   return (
     <div className="flex justify-center min-h-screen bg-gray-100">
@@ -71,9 +48,9 @@ export default function ArticleForm() {
             <label htmlFor="img" className={classNames(styleLabel)}>
               Seleccione imagen
             </label>
-            <ImageUpload onFileChange={handleFileChange} />
+              <ImageUpload onFileChange={handleFileChange} />
           </div>
-          
+
           <button
             type="submit"
             className="self-center mt-6 bg-indigo-500 hover:bg-indigo-700 text-white font-bold py-2 px-6 rounded-lg focus:outline-none focus:shadow-outline"
